+++ docs/changelog.md
@@ -2,6 +2,11 @@
 
 ## Version `1.2.31-ua`
 ### feat
+- #### SKInput
+  - add SKInputVC.Type.VisiblePassword for SkInput
+
+## Version `1.2.31-ua`
+### feat
 - add drawable-nodpi in Icons file generation
 
 ## Version `1.2.30-ua`